Fix broken reference to `fail()` after jest-upgrade
Jest versions >= 27 replace jest-jasmine2 with jest-circus, which does
not have an implementation for the `fail()` method, but still has
the method stub defined in the Jest type-defs. This change implements
the `fail()` method in protocol test suites, and adjusts the thrown
message to actually state the error, instead of just `err`.